add attribute logging to up.rs
