+++ src/util/mod.rs
@@ -1,2 +1,3 @@
+pub mod logs;
 pub mod prompt;
 pub mod tokio_spawner;
